created the development starter project

+++ src/main/java/org/example/ruwaa/Repository/UsersRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsersRepository extends JpaRepository<Users, Long>
+public interface UsersRepository extends JpaRepository<Users, Integer>
 {
     @Query("select u from Users u where u.id = :id")
     Users findUserById(Integer id);
